requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Werkzeug==3.0.3
cryptography>=41.0.0
pytz>=2024.1
reportlab>=4.0.0
python-dotenv>=1.0.0